Move and rename resolve-types file
